requirements.txt
alkana
emoji
keyboard
mecab-python3
openai==1.42.0
pandas
PyAudio
pyautogui
pytchat
requests
sounddevice
soundfile
unidic
httpx==0.25.0
TikTokLive
httpcore==0.18.0
python-dotenv
langchain
langchain-openai
twitchAPI
elevenlabs
